Move the code based on the official Kube client to a separate package
