+++ titus-common-ext/kube/gradle/dependency-locks/testCompileClasspath.lockfile
@@ -3,14 +3,17 @@
 # This file is expected to be part of source control.
 aopalliance:aopalliance:1.0
 cglib:cglib-nodep:3.1
-com.fasterxml.jackson.core:jackson-annotations:2.10.5
-com.fasterxml.jackson.core:jackson-core:2.10.5
-com.fasterxml.jackson.core:jackson-databind:2.10.5.1
+com.fasterxml.jackson.core:jackson-annotations:2.11.2
+com.fasterxml.jackson.core:jackson-core:2.11.2
+com.fasterxml.jackson.core:jackson-databind:2.11.2
+com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.11.2
 com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.10.5
+com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.11.2
 com.fasterxml:classmate:1.3.1
 com.flipkart.zjsonpatch:zjsonpatch:0.4.11
 com.github.akarnokd:rxjava2-interop:0.13.7
 com.github.ben-manes.caffeine:caffeine:2.9.0
+com.github.mifmif:generex:1.0.2
 com.github.vladimir-bukhtoyarov:bucket4j-core:6.0.1
 com.google.api.grpc:proto-google-common-protos:2.0.1
 com.google.code.findbugs:jsr305:3.0.2
@@ -39,6 +42,29 @@ com.squareup.okhttp3:okhttp:3.14.9
 com.squareup.okio:okio:1.17.2
 commons-codec:commons-codec:1.15
 commons-io:commons-io:2.8.0
+dk.brics.automaton:automaton:1.11-8
+io.fabric8:kubernetes-client:5.8.0
+io.fabric8:kubernetes-model-admissionregistration:5.8.0
+io.fabric8:kubernetes-model-apiextensions:5.8.0
+io.fabric8:kubernetes-model-apps:5.8.0
+io.fabric8:kubernetes-model-autoscaling:5.8.0
+io.fabric8:kubernetes-model-batch:5.8.0
+io.fabric8:kubernetes-model-certificates:5.8.0
+io.fabric8:kubernetes-model-common:5.8.0
+io.fabric8:kubernetes-model-coordination:5.8.0
+io.fabric8:kubernetes-model-core:5.8.0
+io.fabric8:kubernetes-model-discovery:5.8.0
+io.fabric8:kubernetes-model-events:5.8.0
+io.fabric8:kubernetes-model-extensions:5.8.0
+io.fabric8:kubernetes-model-flowcontrol:5.8.0
+io.fabric8:kubernetes-model-metrics:5.8.0
+io.fabric8:kubernetes-model-networking:5.8.0
+io.fabric8:kubernetes-model-node:5.8.0
+io.fabric8:kubernetes-model-policy:5.8.0
+io.fabric8:kubernetes-model-rbac:5.8.0
+io.fabric8:kubernetes-model-scheduling:5.8.0
+io.fabric8:kubernetes-model-storageclass:5.8.0
+io.fabric8:zjsonpatch:0.3.0
 io.grpc:grpc-api:1.39.0
 io.grpc:grpc-context:1.39.0
 io.grpc:grpc-core:1.39.0
@@ -110,7 +136,7 @@ org.jboss.logging:jboss-logging:3.3.0.Final
 org.mockito:mockito-core:3.3.3
 org.objenesis:objenesis:2.6
 org.reactivestreams:reactive-streams:1.0.3
-org.slf4j:slf4j-api:1.7.30
+org.slf4j:slf4j-api:1.7.32
 org.slf4j:slf4j-log4j12:1.7.0
 org.springframework.boot:spring-boot-autoconfigure:2.3.6.RELEASE
 org.springframework.boot:spring-boot:2.3.6.RELEASE
