+++ titus-common-ext/kube/src/main/java/com/netflix/titus/ext/kube/clustermembership/connector/transport/main/KubeUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Netflix, Inc.
+ * Copyright 2021 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.ext.kube.clustermembership.connector;
+package com.netflix.titus.ext.kube.clustermembership.connector.transport.main;
 
 import com.netflix.titus.api.clustermembership.connector.ClusterMembershipConnectorException;
 import com.netflix.titus.common.util.ExceptionExt;
